Update MailboxService.php (In testing)
I have had a persistent issue with some emails generating tickets but threads are empty. This causes tickets to be un-assignable and unreadable from the UI (the error in DEV mode materializes as a "createdat" null variable when trying to view a ticket). Going through the code revealed that this exception is not handled. I am not a high level programmer so I am still working on this and testing. Goal is to use existing codebase.